feature: add monitor exam timer, exam duration to int

+++ mesh-exam/android-base/src/main/java/io/flaterlab/meshexam/androidbase/common/adapter/ExamListAdapter.kt
@@ -57,7 +57,7 @@ class ExamListAdapter @Inject constructor(
     interface ExamItem {
         val id: String
         val name: String
-        val durationInMin: Long
+        val durationInMin: Int
     }
 
     companion object {
