+++ mesh-exam/domain/api/src/main/java/io/flaterlab/meshexam/domain/create/model/CreateExamModel.kt
@@ -3,5 +3,5 @@ package io.flaterlab.meshexam.domain.create.model
 data class CreateExamModel(
     val name: String,
     val type: String?,
-    val durationInMin: Long,
+    val durationInMin: Int,
 )
