+++ mesh-exam/feature/create/src/main/java/io/flaterlab/meshexam/create/dvo/QuestionMetaDvo.kt
@@ -3,5 +3,5 @@ package io.flaterlab.meshexam.create.dvo
 data class QuestionMetaDvo(
     val name: String,
     val type: String,
-    val durationInMin: Long,
+    val durationInMin: Int,
 )
